+++ src/Microsoft/Silk.NET.Direct3D12/Structs/RaytracingInstanceDesc.gen.cs
@@ -66,7 +66,7 @@ namespace Silk.NET.Direct3D12
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
             get => (uint)(_bitfield1 & 0xFFFFFFu);
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            set =>_bitfield1 = (uint)((_bitfield1 & ~0xFFFFFFu) | (uint)((uint)(value) & 0xFFFFFFu));
+            set => _bitfield1 = (uint)((uint)(_bitfield1 & ~0xFFFFFFu) | (uint)((uint)(value) & 0xFFFFFFu));
         }
 
         public uint InstanceMask
@@ -74,7 +74,7 @@ namespace Silk.NET.Direct3D12
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
             get => (uint)((_bitfield1 >> 24) & 0xFFu);
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            set =>_bitfield1 = (uint)((_bitfield1 & ~(0xFFu << 24)) | (uint)(((uint)(value) & 0xFFu) << 24));
+            set => _bitfield1 = (uint)((uint)(_bitfield1 & ~(0xFFu << 24)) | (uint)(((uint)(value) & 0xFFu) << 24));
         }
 
         private uint _bitfield2;
@@ -84,7 +84,7 @@ namespace Silk.NET.Direct3D12
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
             get => (uint)(_bitfield2 & 0xFFFFFFu);
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            set =>_bitfield2 = (uint)((_bitfield2 & ~0xFFFFFFu) | (uint)((uint)(value) & 0xFFFFFFu));
+            set => _bitfield2 = (uint)((uint)(_bitfield2 & ~0xFFFFFFu) | (uint)((uint)(value) & 0xFFFFFFu));
         }
 
         public uint Flags
@@ -92,7 +92,7 @@ namespace Silk.NET.Direct3D12
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
             get => (uint)((_bitfield2 >> 24) & 0xFFu);
             [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            set =>_bitfield2 = (uint)((_bitfield2 & ~(0xFFu << 24)) | (uint)(((uint)(value) & 0xFFu) << 24));
+            set => _bitfield2 = (uint)((uint)(_bitfield2 & ~(0xFFu << 24)) | (uint)(((uint)(value) & 0xFFu) << 24));
         }
 
         [NativeName("Type", "D3D12_GPU_VIRTUAL_ADDRESS")]
